Admin用APIのネームスペースを分ける
